+++ asf_search/search/__init__.py
@@ -2,4 +2,4 @@ from .search import search
 from .granule_search import granule_search
 from .product_search import product_search
 from .geo_search import geo_search
-from .baseline_search import stack_from_id
+from .baseline_search import stack_from_id, get_stack_params, calc_temporal_baselines