+++ asf_search/search/granule_search.py
@@ -7,7 +7,7 @@ from asf_search.constants import INTERNAL
 
 def granule_search(
         granule_list: Iterable[str],
-        host: str = INTERNAL.SEARCH_API_HOST,
+        host: str = None,
         cmr_token: str = None,
         cmr_provider: str = None
 ) -> ASFSearchResults: