+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [3.0.1](https://github.com/asfadmin/Discovery-asf_search/compare/v3.0.0...v3.0.1)
+### Added
+- `ASFProduct.nearest_neighbors()` provides `depth` temporally nearest neighbors (preceding) from a product's stack
+- More docstrings
+### Changed
+- `ASFProduct.stack()` and the generic `stack_from_id()` now support `start` and `end` parameters for limiting stacks
+
 ## [3.0.0](https://github.com/asfadmin/Discovery-asf_search/compare/v2.0.2...v3.0.0)
 ### Added
 - Auth support for username/password and cookiejars, in addition to the previously available token-based approach. Create a session, authenticate it with the method of choice, then pass the session to whichever download method is being used.